fixed issue with counter length

+++ fleetapp/ViewModels/TruckTypeMinePlanViewModel .cs
@@ -10,6 +10,7 @@ using System.Windows.Controls;
 using System.Collections.ObjectModel;
 using System.Windows.Data;
 using System.IO;
+using System.Windows;
 
 namespace fleetapp.ViewModels
 {
@@ -39,11 +40,18 @@ namespace fleetapp.ViewModels
 
         public String TruckMinePlanFile
         {
+            get { return _truckMinePlanFileName; }
             set { _truckMinePlanFileName = value; }
         }
 
         public void ImportFile()
         {
+            if (String.IsNullOrEmpty(_truckMinePlanFileName))
+            {
+                Console.WriteLine("Please select a file");
+                MessageBox.Show("Please select a file!");
+                return;
+            }
             IEnumerable<TruckTypeMinePlanModel> newTruckTypeMinePlans = ReadCSV(_truckMinePlanFileName);
             _TruckTypeMinePlanDataAccess.DeleteAll();
             _TruckTypeMinePlanDataAccess.InsertTruckTypeMinePlans(newTruckTypeMinePlans);
@@ -74,11 +82,9 @@ namespace fleetapp.ViewModels
             string[] lines = File.ReadAllLines(System.IO.Path.ChangeExtension(fileName, ".csv"));
             String[] Headers = lines[0].Split(',');
             lines = lines.Skip(1).ToArray();          
-            var _assetNumber = 0;
             return lines.Select(line =>
             {
                 string[] data = line.Split(',');
-                _assetNumber += 1;
                 TruckTypeMinePlanModel newTruckTypeMinePlan = new TruckTypeMinePlanModel
                 {
                     ScenarioId = Context.ScenarioId,
@@ -87,7 +93,7 @@ namespace fleetapp.ViewModels
                     MinePlanPayload = Int32.Parse(data[2])
                 };
                 newTruckTypeMinePlan.TruckTypeMinePlanYearMapping = new List<TruckTypeMinePlanYearMappingModel>();
-                for (var i= 3; i<line.Length; i++)
+                for (var i= 3; i<data.Length; i++)
                 {
                     TruckTypeMinePlanYearMappingModel TruckTypeMinePlanYearMapping = new TruckTypeMinePlanYearMappingModel
                     {
